requirements.txt
fastapi
uvicorn
langchain
pydantic
python-multipart
PyPDF2
python-docx
spacy
transformers
google-generativeai
python-dotenv
scikit-learn
numpy